+++ src/RadixDlt.NetworkGateway.PostgresIntegration/Models/LedgerTransaction.cs
@@ -163,8 +163,17 @@ internal class TransactionReceipt
     [Column("receipt_error_message")]
     public string? ErrorMessage { get; set; }
 
-    [Column("receipt_events", TypeName = "jsonb")]
-    public string? Events { get; set; }
+    [Column("receipt_events_sbor")]
+    public byte[][]? EventsSbor { get; set; }
+
+    [Column("receipt_events_schema_hash")]
+    public byte[][]? EventsSchemaHash { get; set; }
+
+    [Column("receipt_events_type_index")]
+    public int[]? EventsTypeIndex { get; set; }
+
+    [Column("receipt_events_type_kind")]
+    public KeyTypeKind[]? EventsTypeKind { get; set; }
 }
 
 internal class GenesisLedgerTransaction : LedgerTransaction